agent/requirements.txt
pydantic-ai[google]>=0.0.14
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
